refactor: 점수 수정 모달 공통 컴포넌트로 분리
