+++ src/widget/member/ui/AcademicScoreEdit/index.tsx
@@ -2,9 +2,7 @@
 
 import type { Member } from '@/feature/member/model/types';
 import { postAcademicScore } from '@/feature/member/api/postScoresByAcademic';
-import { toast } from 'sonner';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useState } from 'react';
+import ScoreEditModal from '@/widget/member/ui/ScoreEditModal';
 
 interface AcademicScoreProps {
   selectedMember: Member;
@@ -12,75 +10,15 @@ interface AcademicScoreProps {
   onSaveSuccess: () => void;
 }
 
-export default function AcademicScoreEdit({
-  selectedMember,
-  onBack,
-  onSaveSuccess,
-}: AcademicScoreProps) {
-  const [value, setValue] = useState('');
-  const queryClient = useQueryClient();
-
-  const { mutate, isPending } = useMutation({
-    mutationFn: postAcademicScore,
-    onSuccess: (_data, variables) => {
-      const { memberId } = variables;
-      queryClient.invalidateQueries({
-        queryKey: ['totalScore', memberId],
-      });
-      queryClient.invalidateQueries({
-        queryKey: ['scoresByCategory', memberId],
-      });
-      onSaveSuccess();
-    },
-    onError: () => {
-      toast.error('유효하지 않은 점수 값입니다.');
-    },
-  });
-
-  const handleSubmit = () => {
-    if (!value) {
-      toast.warning('교과 성적 평균 등급을 입력해주세요.');
-      return;
-    }
-    mutate({
-      value,
-      memberId: selectedMember.id,
-    });
-  };
-
+export default function AcademicScoreEdit(props: AcademicScoreProps) {
   return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/20">
-      <div className="h-[361px] w-[500px] rounded-[20px] bg-white px-[60px] py-[40px]">
-        <p className="text-main-700 mb-[24px] text-2xl font-semibold">교과성적 점수 변경</p>
-
-        <div className="mb-[30px] flex flex-col gap-1">
-          <p>교과 성적 입력</p>
-          <input
-            className="h-[56px] w-[380px] rounded-xl border border-gray-300 px-[16px] text-sm font-medium"
-            type="number"
-            placeholder="교과 성적 평균 등급 입력"
-            value={value}
-            onChange={(e) => setValue(e.target.value)}
-          />
-        </div>
-
-        <div className="flex flex-col gap-3">
-          <button
-            onClick={onBack}
-            className="border-main-500 text-main-500 h-[52px] w-[380px] rounded-xl border text-lg font-semibold"
-          >
-            뒤로가기
-          </button>
-
-          <button
-            onClick={handleSubmit}
-            disabled={isPending}
-            className="bg-main-500 h-[52px] w-[380px] rounded-xl text-lg font-semibold text-white disabled:opacity-50"
-          >
-            변경 완료
-          </button>
-        </div>
-      </div>
-    </div>
+    <ScoreEditModal
+      {...props}
+      title="교과성적 점수 변경"
+      label="교과 성적 입력"
+      placeholder="교과 성적 평균 등급 입력"
+      warningMessage="교과 성적 평균 등급을 입력해주세요."
+      mutationFn={postAcademicScore}
+    />
   );
 }
