+++ src/widget/member/ui/VolunteerScoreEdit/index.tsx
@@ -2,9 +2,7 @@
 
 import type { Member } from '@/feature/member/model/types';
 import { postVolunteerScore } from '@/feature/member/api/postScoresByVolunteer';
-import { toast } from 'sonner';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useState } from 'react';
+import ScoreEditModal from '@/widget/member/ui/ScoreEditModal';
 
 interface VolunteerScoreProps {
   selectedMember: Member;
@@ -12,75 +10,15 @@ interface VolunteerScoreProps {
   onSaveSuccess: () => void;
 }
 
-export default function VolunteerScoreEdit({
-  selectedMember,
-  onBack,
-  onSaveSuccess,
-}: VolunteerScoreProps) {
-  const [value, setValue] = useState('');
-  const queryClient = useQueryClient();
-
-  const { mutate, isPending } = useMutation({
-    mutationFn: postVolunteerScore,
-    onSuccess: (_data, variables) => {
-      const { memberId } = variables;
-      queryClient.invalidateQueries({
-        queryKey: ['totalScore', memberId],
-      });
-      queryClient.invalidateQueries({
-        queryKey: ['scoresByCategory', memberId],
-      });
-      onSaveSuccess();
-    },
-    onError: () => {
-      toast.error('유효하지 않은 점수 값입니다.');
-    },
-  });
-
-  const handleSubmit = () => {
-    if (!value) {
-      toast.warning('총 봉사 시간을 입력해주세요.');
-      return;
-    }
-    mutate({
-      value,
-      memberId: selectedMember.id,
-    });
-  };
-
+export default function VolunteerScoreEdit(props: VolunteerScoreProps) {
   return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black/20">
-      <div className="h-[361px] w-[500px] rounded-[20px] bg-white px-[60px] py-[40px]">
-        <p className="text-main-700 mb-[24px] text-2xl font-semibold">봉사 점수 변경</p>
-
-        <div className="mb-[30px] flex flex-col gap-1">
-          <p>총 봉사 시간 입력</p>
-          <input
-            className="h-[56px] w-[380px] rounded-xl border border-gray-300 px-[16px] text-sm font-medium"
-            type="number"
-            placeholder="교육과정 봉사활동 제외, 총 봉사 시간 입력"
-            value={value}
-            onChange={(e) => setValue(e.target.value)}
-          />
-        </div>
-
-        <div className="flex flex-col gap-3">
-          <button
-            onClick={onBack}
-            className="border-main-500 text-main-500 h-[52px] w-[380px] rounded-xl border text-lg font-semibold"
-          >
-            뒤로가기
-          </button>
-
-          <button
-            onClick={handleSubmit}
-            disabled={isPending}
-            className="bg-main-500 h-[52px] w-[380px] rounded-xl text-lg font-semibold text-white disabled:opacity-50"
-          >
-            변경 완료
-          </button>
-        </div>
-      </div>
-    </div>
+    <ScoreEditModal
+      {...props}
+      title="봉사 점수 변경"
+      label="총 봉사 시간 입력"
+      placeholder="교육과정 봉사활동 제외, 총 봉사 시간 입력"
+      warningMessage="총 봉사 시간을 입력해주세요."
+      mutationFn={postVolunteerScore}
+    />
   );
 }
